Refactor project structure and add prop types
Changes include:
- Moved components into a 'components' directory
- Installed prop-types via yarn
- Defined propTypes for all components
- Ensured app renders without console errors
