+++ react-todo/src/components/InputWithLabel.jsx
@@ -1,4 +1,5 @@
 import React, { useRef, useEffect } from "react";
+import PropTypes from "prop-types";
 
 
 function InputWithLabel({ id, value, type = "text", onInputChange, children }) {
@@ -21,5 +22,12 @@ function InputWithLabel({ id, value, type = "text", onInputChange, children }) {
         </>
     );
 }
+
+InputWithLabel.propTypes = {
+    id: PropTypes.string.isRequired, 
+    value: PropTypes.string.isRequired,
+    onInputChange: PropTypes.func.isRequired,
+    type: PropTypes.string, 
+};
 
 export default InputWithLabel
