+++ react-todo/src/components/TodoList.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import TodoListItem from "./TodoListItem";
+import PropTypes from "prop-types";
 
 function TodoList({ todoList, onRemoveTodo }) {
   return (
@@ -10,5 +11,15 @@ function TodoList({ todoList, onRemoveTodo }) {
     </ul>
   );
 }
+
+TodoList.propTypes = {
+  todoList: PropTypes.arrayOf(
+      PropTypes.shape({
+          id: PropTypes.string.isRequired,
+          title: PropTypes.string.isRequired,
+      })
+  ).isRequired,
+  onRemoveTodo: PropTypes.func.isRequired,
+};
 
 export default TodoList;
