+++ react-todo/src/index.html
@@ -11,5 +11,6 @@
   </head>
   <body>
     <div id="root"></div>
+    <!-- <script type="module" src="/src/main.tsx"></script> -->
   </body>
 </html>
