+++ react-todo/src/main.jsx
@@ -1,6 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
+import App from './App.jsx';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
